+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.8.0] - 2025-05-22
+### Added
+- [#5] Add new field for minimal data volume size
+### Changed
+- [#5] mark dataVolumeSize as deprecated in favour of minDataVolumeSize
+
 ## [v2.7.0] - 2025-05-20
 ### Added
 - [#3] Add datamount spec to dogu crd. This adds the possibility to mount configmaps & secrets directly to dogu.json-Volumes.